make PUT request from client to update posters
