+++ src/components/TableSection/MoviePoster.component.tsx
@@ -1,6 +1,5 @@
 import Skeleton from "react-loading-skeleton";
 import "react-loading-skeleton/dist/skeleton.css";
-import { useResize } from "../../hooks/useResize.hook";
 import { useContext } from "react";
 import { MoviesContext } from "../../contexts/movies.context";
 import { MoviesContextType } from "../../types";
